move to util.go
